Changing history API fallback to Express middleware
The previous Connect middleware rewrites whether or not the file
exists. This causes issues with paths containing dots and paths without
extensions. The Express middleware is simpler. It is called only if the
file is not found. And it can be passed an absolute file path. E.g.
`__dirname + '/public/index.html'`